ref: timetable also fix up eventFiltersModal story
